requirements.txt
google-genai>=1.0.0
python-docx>=1.1.0
python-dotenv>=1.0.0
pyyaml>=6.0
streamlit>=1.30.0
